Fixed timing issues in main, time based movement
Only implemented time based movement for player so far, needs to be
implemented for other players, monsters and probably projectiles as
well.  Also, if you want it to be optional, you'll need to add an option
and wrap the new and old in an conditional.  The part that updates cwalk
should be good for both (D will need to get assigned outside of the
conditional).

Also, time issues in main can be fixed independently of this (the
changes in sub main()).  This is what is causing 39 fps instead of 40.